feat: rewrite to modern svelte kit
