+++ .gitignore
@@ -1,5 +1,28 @@
-node_modules/
-dist/
-src/config.ts
-*lock.*
-!package-lock.json
+node_modules
+
+# Output
+.output
+.vercel
+.netlify
+.wrangler
+/.svelte-kit
+/build
+
+# OS
+.DS_Store
+Thumbs.db
+
+# Env
+.env
+.env.*
+!.env.example
+!.env.test
+
+# Vite
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
+
+# Config
+/src/lib/config.ts
+
+bun.lock
