--- 404.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<base href="/" />
-		<meta charset="UTF-8" />
-		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-		<title>KekUpload</title>
-		<meta name="title" content="KekUpload" />
-		<meta property="og:title" content="KekUpload" />
-		<meta name="description" content="Upload everything" />
-		<meta property="og:description" content="Upload everything" />
-		<meta property="twitter:description" content="Upload everything" />
-		<meta name="theme-color" content="#000000" />
-
-		<link rel="stylesheet" id="theme" href="/themes/korng.css" />
-		<link rel="stylesheet" href="/404.css" />
-	</head>
-	<body>
-		<h1 id="title">404</h1>
-		<p id="description">This page could not be found</p>
-		<a id="home-link" href="/">Back to uploads</a>
-		<script type="module" src="/src/404.ts"></script>
-	</body>
-</html>
