--- index.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta base="/" />
-
-		<link rel="icon" href="/favicon.ico" />
-		<link rel="manifest" href="manifest.json" />
-
-		<meta charset="UTF-8" />
-
-		<!-- CONFIG -->
-		<!-- Enter your own title -->
-		<title>KekUpload</title>
-		<meta name="title" content="KekUpload" />
-		<meta property="og:title" content="KekUpload" />
-
-		<!-- Enter your own description -->
-		<meta name="description" content="Upload everything" />
-		<meta property="og:description" content="Upload everything" />
-		<meta property="twitter:description" content="Upload everything" />
-
-		<!-- Enter your own theme color -->
-		<meta name="theme-color" content="#000000" />
-		<!-- CONFIG -->
-
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-		<link rel="stylesheet" id="theme" href="/themes/korng.css" />
-	</head>
-	<body>
-		<div id="app"></div>
-		<script type="module" src="/src/main.ts"></script>
-	</body>
-</html>
