+++ src/lib/components/UploadList.svelte
@@ -1,0 +1,57 @@
+<script lang="ts">
+	import UploadEntry from "./UploadEntry.svelte";
+
+	let {
+		onfileschange
+	}: {
+		onfileschange?: (count: number) => void;
+	} = $props();
+
+	let uploadEntries: Record<number, Record<string, any>> = {};
+	let files: Array<File> = $state([]);
+
+	function removeitem(index: number) {
+		files.splice(index, 1);
+		files = files;
+		onfileschange?.(files.length);
+	}
+
+	function uploadAll() {
+		Object.values(uploadEntries).forEach((entry) => {
+			entry.startUpload();
+		});
+	}
+
+	function addFiles(filelist: FileList) {
+		for (let file of filelist) files.push(file);
+
+		files = files;
+		onfileschange?.(files.length);
+	}
+</script>
+
+{#if files.length > 0}
+	<main>
+		{#each files as file, i (file)}
+			<UploadEntry
+				bind:this={uploadEntries[i]}
+				{file}
+				index={i}
+				onclose={() => removeitem(i)}
+			/>
+			<div class="spacer"></div>
+		{/each}
+	</main>
+{/if}
+
+<style>
+	main {
+		width: 90%;
+		border: 1px solid var(--border);
+		border-radius: 15px;
+		padding: 10px 5px 0 5px;
+	}
+	.spacer {
+		margin-bottom: 10px;
+	}
+</style>
