+++ static/manifest.json
