+++ static/robots.txt
@@ -1,0 +1,3 @@
+# allow crawling everything by default
+User-agent: *
+Disallow:
