--- svelte.config.js
@@ -1,7 +1,0 @@
-import sveltePreprocess from "svelte-preprocess";
-
-export default {
-	// Consult https://github.com/sveltejs/svelte-preprocess
-	// for more information about preprocessors
-	preprocess: sveltePreprocess()
-};
