--- tsconfig.node.json
@@ -1,8 +1,0 @@
-{
-	"compilerOptions": {
-		"composite": true,
-		"module": "esnext",
-		"moduleResolution": "node"
-	},
-	"include": ["vite.config.ts"]
-}
