+++ vite.config.ts
@@ -1,22 +1,20 @@
-import { defineConfig } from "vite";
-import { svelte } from "@sveltejs/vite-plugin-svelte";
-import legacy from "@vitejs/plugin-legacy";
+import adapter from '@sveltejs/adapter-auto';
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig } from 'vite';
 
-// https://vitejs.dev/config/
 export default defineConfig({
-	build: {
-		rollupOptions: {
-			input: {
-				main: "index.html",
-				notFound: "404.html"
-			}
-		}
-	},
 	plugins: [
-		svelte(),
-		legacy({
-			targets: ["ie >= 11"],
-			additionalLegacyPolyfills: ["regenerator-runtime/runtime"]
+		sveltekit({
+			compilerOptions: {
+				// Force runes mode for the project, except for libraries. Can be removed in svelte 6.
+				runes: ({ filename }) =>
+					filename.split(/[/\\]/).includes('node_modules') ? undefined : true
+			},
+
+			// adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
+			// If your environment is not supported, or you settled on a specific environment, switch out the adapter.
+			// See https://svelte.dev/docs/kit/adapters for more information about adapters.
+			adapter: adapter()
 		})
 	]
 });
